requirements.txt
django==5.2.7
langchain==0.3.27
python-dotenv
pypdf==6.1.1
langchain==0.3.27
langgraph==0.6.2
langchain-chroma==0.2.5
langgraph-checkpoint-sqlite==2.0.11
langchain-community==0.3.27
langchain-core==0.3.72
python-dotenv==1.1.1
djangorestframework==3.16.1
langchain-together==0.3.1
langchain-openai==0.3.28
llama_parse==0.6.56
